pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "homie-logging"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yaqc",
    "scipy",
    "numpy",
    "Homie4",
]
version = "0.1.0"

[project.scripts]
mqtt-publish = "mqtt.main:main"
show-spectrum = "gui.spec:main"
